refactor: replace files to another folders
