+++ src/app/App.js
@@ -1,7 +1,7 @@
 import { homedir } from "node:os";
-import { welcomeMessage} from '../components/general.js';
-import { getDirectoryPath } from '../utils/get-directory-path.js';
-import { beginCLI } from "../components/begin-cli.js";
+import { welcomeMessage } from "../components/base/general.js";
+import { getDirectoryPath } from "../utils/get-directory-path.js";
+import { beginCLI } from "../components/base/begin-cli.js";
 
 export class App {
   async start() {
