--- src/components/go-to-folder.js
@@ -1,7 +1,0 @@
-  import { getAbsolutePath } from "../utils/get-absolute-path.js";
-
-export const goToFolder = async ([...args]) => {
-  process.chdir(getAbsolutePath(args));
-};
-
-
