+++ src/utils/go-to-folder.js
@@ -1,0 +1,5 @@
+import { getAbsolutePath } from './get-absolute-path.js';
+
+export const goToFolder = async ([...args]) => {
+    process.chdir(getAbsolutePath(args));
+};
